Split catalog module
This commit brings in a new datastore module that replaces old catalog,
and contains all datastore container modules.
The split should improve maintainability of these modules.
