+++ hyena-engine/src/datastore/mod.rs
@@ -2,7 +2,6 @@ use error::*;
 use ty::{BlockType as TyBlockType, ColumnId};
 use hyena_common::ty::{Timestamp, MIN_TIMESTAMP};
 use block::{BlockType, SparseIndex};
-use partition::{Partition, PartitionId};
 use storage::manager::{PartitionGroupManager, PartitionManager};
 use std::collections::hash_map::HashMap;
 use std::collections::vec_deque::VecDeque;
@@ -19,732 +18,21 @@ use scanner::{Scan, ScanResult};
 use ty::block::{BlockTypeMap, BlockTypeMapTy};
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
 
+mod catalog;
+mod column;
+mod partition;
+mod partition_group;
+mod partition_meta;
+
+pub use self::catalog::Catalog;
+pub use self::column::Column;
+pub(crate) use self::partition::{Partition, PartitionId};
+pub(crate) use self::partition_meta::PartitionMeta;
+pub(crate) use self::partition_group::PartitionGroup;
+
 pub(crate) type PartitionMap<'part> = HashMap<PartitionMeta, Partition<'part>>;
 pub(crate) type PartitionGroupMap<'pg> = HashMap<SourceId, PartitionGroup<'pg>>;
 pub type ColumnMap = HashMap<ColumnId, Column>;
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Catalog<'cat> {
-    pub(crate) columns: ColumnMap,
-
-    pub(crate) groups: PartitionGroupMap<'cat>,
-
-    #[serde(skip)]
-    pub(crate) data_root: PathBuf,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub(crate) struct PartitionGroup<'pg> {
-    // TODO: we should consider changing this to something more universal
-    // and less coupled with our specific schema perhaps
-    source_id: SourceId,
-
-    #[serde(skip)]
-    pub(crate) immutable_partitions: PartitionMap<'pg>,
-
-    #[serde(skip)]
-    pub(crate) mutable_partitions: RwLock<VecDeque<Partition<'pg>>>,
-
-    #[serde(skip)]
-    data_root: PathBuf,
-}
-
-impl<'pg> PartitionGroup<'pg> {
-    fn new<P: AsRef<Path>>(root: P, source_id: SourceId) -> Result<PartitionGroup<'pg>> {
-        let root = root.as_ref().to_path_buf();
-
-        let meta = root.join(PARTITION_GROUP_METADATA);
-
-        if meta.exists() {
-            bail!("Partition group metadata already exists {:?}", meta);
-        }
-
-        Ok(PartitionGroup {
-            source_id,
-            immutable_partitions: Default::default(),
-            mutable_partitions: Default::default(),
-            data_root: root,
-        })
-    }
-
-    pub fn with_data<P: AsRef<Path>>(root: P) -> Result<PartitionGroup<'pg>> {
-        let root = root.as_ref().to_path_buf();
-
-        let meta = root.join(PARTITION_GROUP_METADATA);
-
-        PartitionGroup::deserialize(&meta, &root)
-    }
-
-    fn flush(&self) -> Result<()> {
-        // TODO: add dirty flag
-        let meta = self.data_root.join(PARTITION_GROUP_METADATA);
-
-        for p in self.immutable_partitions.values() {
-            p.flush()?
-        }
-
-        for p in self.mutable_partitions.read().unwrap().iter() {
-            p.flush()?
-        }
-
-        PartitionGroup::serialize(self, &meta)
-    }
-
-    fn append(&self, catalog: &Catalog, data: &Append) -> Result<usize> {
-        use std::iter::{once, repeat};
-        use ty::fragment::FragmentRef;
-
-        // calculate the size of the append (in records)
-
-        let mut colindices = vec![0, 1]; // ts and source_id
-        colindices.extend(data.data.keys());
-
-        let columns = catalog.as_ref();
-
-        let typemap: BlockTypeMap = colindices
-            .iter()
-            .map(|colidx| if let Some(col) = columns.get(colidx) {
-                Ok((*colidx, **col))
-            } else {
-                bail!("column {} not found", colidx)
-            })
-            .collect::<Result<BlockTypeMapTy>>()
-            .with_context(|_| err_msg("failed to prepare all columns"))?
-            .into();
-
-        let fragcount = data.len();
-
-        let (emptycap, currentcap) = {
-            let partitions = acquire!(read carry self.mutable_partitions);
-            // current partition capacity
-            let curpart = partitions
-                .back()
-                .ok_or_else(|| err_msg("Mutable partitions pool is empty"))?;
-
-            // empty partition capacity for this append
-            let emptycap = catalog.space_for_blocks(&colindices);
-            let currentcap = curpart.space_for_blocks(&colindices);
-
-            (
-                emptycap,
-                // check if current partition exceeded its capacity or is simply uninitialized
-                if currentcap == 0 && curpart.is_empty() {
-                    emptycap
-                } else {
-                    currentcap
-                },
-            )
-        };
-
-        // check if we can fit the data within current partition
-        // or additional ones are needed
-
-        let (curfrags, reqparts, current_is_full) = if fragcount > currentcap {
-            let emptyfrags = fragcount - currentcap;
-            let lastfull = emptyfrags % emptycap == 0;
-
-            (
-                currentcap,
-                (emptyfrags / emptycap + if lastfull { 0 } else { 1 }) - 1,
-                lastfull,
-            )
-        } else {
-            (fragcount, 0, fragcount == currentcap)
-        };
-
-        // prepare source slices
-
-        let mut fragments = Vec::with_capacity(reqparts + 1);
-
-        let (fragments, ts_1, mut frag_1, mut ts_idx, offsets) = once(curfrags)
-            .filter(|c| *c != 0)
-            .chain(repeat(emptycap).take(reqparts))
-            .fold(
-                (
-                    &mut fragments,
-                    data.ts.as_slice(),
-                    data.data
-                        .iter()
-                        .map(|(col_id, frag)| (*col_id, FragmentRef::from(frag)))
-                        .collect::<HashMap<_, _>>(),
-                    vec![],
-                    vec![0],
-                ),
-                |store, mid| {
-
-                    let (fragments, ts_data, frag_data, mut ts_idx, mut offsets) = store;
-
-                    // when dealing with sparse blocks we don't know beforehand which
-                    // partition a sparse entry will belong to
-                    // but as our splitting algorithm calculates the capacity with an assumption
-                    // that sparse blocks are in fact dense, the "overflow" of sparse data
-                    // shouldn't happen
-
-                    let (ts_0, ts_1) = ts_data.split_at(mid);
-                    let (mut frag_0, frag_1) = frag_data.iter().fold(
-                        (HashMap::new(), HashMap::new()),
-                        |acc, (col_id, frag)| {
-                            let (mut hm_0, mut hm_1) = acc;
-
-                            if frag.is_sparse() {
-                                // this uses unsafe conversion usize -> u32
-                                // in reality we shouldn't ever use mid > u32::MAX
-                                // it's still worth to consider adding some check
-                                let (frag_0, frag_1) = frag.split_at_idx(mid as SparseIndex)
-                                    .with_context(|_| "unable to split sparse fragment")
-                                    .unwrap();
-
-                                hm_0.insert(*col_id, frag_0);
-                                hm_1.insert(*col_id, frag_1);
-                            } else {
-                                let (frag_0, frag_1) = frag.split_at(mid);
-
-                                hm_0.insert(*col_id, frag_0);
-                                hm_1.insert(*col_id, frag_1);
-                            }
-
-                            (hm_0, hm_1)
-                        },
-                    );
-
-                    // as we did split, ts_0 cannot be empty
-                    ts_idx.push(
-                        *(&ts_0[..]
-                            .first()
-                            .ok_or_else(|| err_msg("ts_0 was empty, this shouldn't happen"))
-                            .unwrap()),
-                    );
-                    frag_0.insert(0, FragmentRef::from(&ts_0[..]));
-                    fragments.push(frag_0);
-                    offsets.push(mid);
-
-                    (fragments, ts_1, frag_1, ts_idx, offsets)
-                },
-            );
-
-        if !ts_1.is_empty() {
-            ts_idx.push(*(&ts_1[..].first().unwrap()));
-            frag_1.insert(0, FragmentRef::from(&ts_1[..]));
-            fragments.push(frag_1);
-        }
-
-        // create partition pool
-
-        let partitions = acquire!(write carry self.mutable_partitions);
-
-        let curidx = partitions.len();
-
-        let newparts = ts_idx
-            .iter()
-            .skip(if current_is_full { 0 } else { 1 })
-            .map(|ts| self.create_partition(**ts))
-            .collect::<Result<Vec<_>>>()
-            .with_context(|_| "Unable to create partition for writing")?;
-
-        partitions.extend(newparts);
-
-        // write data
-
-        for ((partition, fragment), offset) in partitions
-            .iter_mut()
-            .skip(curidx - if current_is_full { 0 } else { 1 })
-            .zip(fragments.iter())
-            .zip(offsets.iter())
-        {
-            partition
-                .append(&typemap, &fragment, *offset as SparseIndex)
-                .with_context(|_| "partition append failed")
-                .unwrap();
-        }
-
-        // todo: this is not very accurate, as the actual write could have failed
-        Ok(fragcount)
-    }
-
-    pub fn scan(&self, scan: &Scan) -> Result<ScanResult> {
-        // only filters and projection for now
-        // all partitions
-        // full ts range
-
-        let partitions = acquire!(read carry self.mutable_partitions);
-
-        partitions
-            .par_iter()
-            .filter(|partition| if let Some(ref scan_partitions) = scan.partitions {
-                // todo: benchmark Partition::get_id() -> &PartitionId
-                scan_partitions.contains(&partition.get_id())
-            } else {
-                true
-            })
-            .map(|partition| {
-                partition
-                    .scan(&scan)
-                    .with_context(|_| "partition scan failed")
-                    .ok()
-            })
-            .reduce(
-                || None,
-                |a, b| if a.is_none() {
-                    b
-                } else if b.is_none() {
-                    a
-                } else {
-                    let mut a = a.unwrap();
-                    let b = b.unwrap();
-                    a.merge(b).unwrap();
-                    Some(a)
-                },
-            )
-            .ok_or_else(|| err_msg("partition scan failed"))
-    }
-
-    fn create_partition<'part, TS>(&self, ts: TS) -> Result<Partition<'part>>
-    where
-        TS: Into<Timestamp> + Clone + Copy,
-    {
-        let part_id = Partition::gen_id();
-
-        let part_root = PartitionManager::new(&self.data_root, part_id, ts)
-            .with_context(|_| "Unable to create partition data path")?;
-
-        Partition::new(part_root, part_id, ts)
-    }
-
-    fn prepare_partitions<P, I>(root: P, ids: I) -> Result<PartitionMap<'pg>>
-    where
-        P: AsRef<Path>,
-        I: IntoIterator<Item = PartitionMeta>,
-    {
-        ids.into_iter()
-            .map(|part_meta| {
-                let path = PartitionManager::new(&root, part_meta.id, part_meta.ts_min)
-                    .with_context(|_| {
-                        format!(
-                            "Unable to obtain data directory for partition {}",
-                            &*part_meta
-                        )
-                    })?;
-
-                let partition = Partition::with_data(path)
-                    .with_context(|_| format!("Unable to read partition {:?}", &*part_meta))?;
-
-                Ok((part_meta, partition))
-            })
-            .collect()
-    }
-
-    fn prepare_mut_partitions<P, I>(root: P, ids: I) -> Result<VecDeque<Partition<'pg>>>
-    where
-        P: AsRef<Path>,
-        I: IntoIterator<Item = PartitionMeta>,
-    {
-        ids.into_iter()
-            .map(|part_meta| {
-                let path = PartitionManager::new(&root, part_meta.id, part_meta.ts_min)
-                    .with_context(|_| {
-                        format!(
-                            "Unable to obtain data directory for partition {}",
-                            &*part_meta
-                        )
-                    })?;
-
-                let partition = Partition::with_data(path)
-                    .with_context(|_| format!("Unable to read partition {:?}", &*part_meta))?;
-
-                Ok(partition)
-            })
-            .collect()
-    }
-
-    fn serialize<P: AsRef<Path>>(pg: &PartitionGroup<'pg>, meta: P) -> Result<()> {
-        let meta = meta.as_ref();
-
-        let im_partition_metas = Vec::from_iter(pg.immutable_partitions.keys());
-        let mut_partition_metas = acquire!(write carry pg.mutable_partitions)
-            .iter()
-            .map(PartitionMeta::from)
-            .collect::<Vec<_>>();
-
-        let data = (pg, im_partition_metas, mut_partition_metas);
-
-        serialize!(file meta, &data)
-            .with_context(|_| "Failed to serialize partition group metadata")
-            .map_err(|e| e.into())
-    }
-
-    fn deserialize<P: AsRef<Path>, R: AsRef<Path>>(
-        meta: P,
-        root: R,
-    ) -> Result<PartitionGroup<'pg>> {
-        let meta = meta.as_ref();
-
-        if !meta.exists() {
-            bail!("Cannot find partition group metadata {:?}", meta);
-        }
-
-        let (mut pg, im_partition_metas, mut_partition_metas): (
-            PartitionGroup,
-            Vec<PartitionMeta>,
-            Vec<PartitionMeta>,
-        ) = deserialize!(file meta)
-            .with_context(|_| "Failed to read catalog metadata")?;
-
-        pg.immutable_partitions = PartitionGroup::prepare_partitions(&root, im_partition_metas)
-            .with_context(|_| "Failed to read immutable partitions data")?;
-
-        pg.mutable_partitions = locked!(rw PartitionGroup::prepare_mut_partitions(
-            &root, mut_partition_metas)
-            .with_context(|_| "Failed to read mutable partitions data")?);
-
-        pg.data_root = root.as_ref().to_path_buf();
-
-        Ok(pg)
-    }
-}
-
-impl<'pg> Drop for PartitionGroup<'pg> {
-    fn drop(&mut self) {
-        self.flush()
-            .with_context(|_| "Failed to flush data during drop")
-            .unwrap();
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Column {
-    pub(crate) ty: TyBlockType,
-    pub(crate) name: String,
-}
-
-impl Column {
-    pub fn new(ty: TyBlockType, name: &str) -> Column {
-        Column {
-            ty,
-            name: name.to_owned(),
-        }
-    }
-}
-
-impl Deref for Column {
-    type Target = TyBlockType;
-
-    fn deref(&self) -> &Self::Target {
-        &self.ty
-    }
-}
-
-impl Display for Column {
-    fn fmt(&self, fmt: &mut Formatter) -> StdResult<(), FmtError> {
-        write!(fmt, "{}", self.name)
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Eq, Ord, Serialize, Deserialize, Hash)]
-pub(crate) struct PartitionMeta {
-    id: PartitionId,
-
-    ts_min: Timestamp,
-    ts_max: Timestamp,
-}
-
-impl PartitionMeta {
-    #[allow(unused)]
-    fn new<TS>(id: PartitionId, ts_min: TS, ts_max: TS) -> PartitionMeta
-    where
-        Timestamp: From<TS>,
-    {
-        PartitionMeta {
-            id,
-            ts_min: ts_min.into(),
-            ts_max: ts_max.into(),
-        }
-    }
-}
-
-impl<'a, 'part> From<&'a Partition<'part>> for PartitionMeta {
-    fn from(partition: &'a Partition<'part>) -> PartitionMeta {
-        let (ts_min, ts_max) = partition.get_ts();
-
-        PartitionMeta {
-            id: partition.get_id(),
-            ts_min,
-            ts_max,
-        }
-    }
-}
-
-impl Deref for PartitionMeta {
-    type Target = PartitionId;
-
-    fn deref(&self) -> &Self::Target {
-        &self.id
-    }
-}
-
-impl<'cat> Catalog<'cat> {
-    pub fn new<P: AsRef<Path>>(root: P) -> Result<Catalog<'cat>> {
-        let root = root.as_ref().to_path_buf();
-
-        let meta = root.join(CATALOG_METADATA);
-
-        if meta.exists() {
-            bail!("Catalog metadata already exists {:?}", meta);
-        }
-
-        let mut catalog = Catalog {
-            columns: Default::default(),
-            groups: Default::default(),
-            data_root: root,
-        };
-
-        catalog.ensure_default_columns()?;
-
-        Ok(catalog)
-    }
-
-    fn ensure_default_columns(&mut self) -> Result<()> {
-        let ts_column = Column::new(TyBlockType::Memmap(BlockType::U64Dense), "timestamp");
-        let source_column = Column::new(TyBlockType::Memory(BlockType::I32Dense), "source_id");
-        let mut map = HashMap::new();
-        map.insert(0, ts_column);
-        map.insert(1, source_column);
-
-        self.ensure_columns(map)
-    }
-
-    pub fn with_data<P: AsRef<Path>>(root: P) -> Result<Catalog<'cat>> {
-        let root = root.as_ref().to_path_buf();
-
-        let meta = root.join(CATALOG_METADATA);
-
-        Catalog::deserialize(&meta, &root)
-    }
-
-    pub fn open_or_create<P: AsRef<Path>>(root: P) -> Result<Catalog<'cat>> {
-        Catalog::with_data(root.as_ref()).or_else(|_| Catalog::new(root.as_ref()))
-    }
-
-    pub fn append(&self, data: &Append) -> Result<usize> {
-        if data.is_empty() {
-            bail!("Provided Append contains no data");
-        }
-
-        // dispatch to proper PartitionGroup
-        if let Some(pg) = self.groups.get(&data.source_id) {
-            pg.append(&self, &data)
-        } else {
-            bail!("No PartitionGroup found for source_id = {}", data.source_id);
-        }
-    }
-
-    pub fn scan(&self, scan: &Scan) -> Result<ScanResult> {
-
-        let all_groups = if scan.groups.is_some() {
-            None
-        } else {
-            Some(self.groups.keys().cloned().collect::<Vec<_>>())
-        };
-
-        if scan.groups.is_some() {
-            scan.groups.as_ref().unwrap()
-        } else {
-            all_groups.as_ref().unwrap()
-        }
-        .par_iter()
-        .filter_map(|pgid| self.groups.get(pgid))
-        .map(|pg| pg.scan(&scan))
-        // todo: this would potentially be better with some short-circuiting combinator instead
-        // need to bench with collect_into()
-        .reduce(|| Ok(ScanResult::merge_identity()), |a, b| {
-            let mut a = a?;
-            let b = b?;
-
-            a.merge(b)?;
-
-            Ok(a)
-        })
-    }
-
-    pub fn flush(&self) -> Result<()> {
-        // TODO: add dirty flag
-        let meta = self.data_root.join(CATALOG_METADATA);
-
-        for pg in self.groups.values() {
-            pg.flush()?
-        }
-
-        Catalog::serialize(self, &meta)
-    }
-
-    /// Extend internal column map without any sanitization checks.
-    ///
-    /// This function uses `std::iter::Extend` internally,
-    /// so it allows redefinition of a column type.
-    /// Use this feature with great caution.
-    pub(crate) fn ensure_columns(&mut self, type_map: ColumnMap) -> Result<()> {
-        self.columns.extend(type_map);
-
-        Ok(())
-    }
-
-    /// Adds the column to the catalog. It verifies that catalog does not already contain:
-    /// a) column with the given id, or
-    /// b) column with the given name.
-    /// This function takes all-or-nothing approach:
-    /// either all columns can are added, or none gets added.
-    pub fn add_columns(&mut self, column_map: ColumnMap) -> Result<()> {
-        for (id, column) in column_map.iter() {
-            info!("Adding column {}:{:?} with id {}", column.name, column.ty, id);
-
-            if self.columns.contains_key(id) {
-                bail!("Column Id already exists {}", *id);
-            }
-            if self.columns.values().any(|col| col.name == column.name) {
-                bail!("Column Name already exists {}", *id);
-            }
-        }
-
-        self.ensure_columns(column_map)
-    }
-
-    /// Fetch the first non-occupied column index
-    ///
-    /// todo: rethink this approach (max() every time)
-    pub fn next_id(&self) -> usize {
-        let default = 0;
-        *self.columns.keys().max().unwrap_or(&default) + 1
-    }
-
-    /// Calculate an empty partition's capacity for given column set
-    fn space_for_blocks(&self, indices: &[usize]) -> usize {
-        use params::BLOCK_SIZE;
-
-        indices.iter()
-            .filter_map(|col_id| {
-                if let Some(column) = self.columns.get(col_id) {
-                    Some(BLOCK_SIZE / column.size_of())
-                } else {
-                    None
-                }
-            })
-            .min()
-            // the default shouldn't ever happen, as there always should be ts block
-            // but in case it happens, this will return 0
-            // which in turn will cause new partition to be used
-            .unwrap_or_default()
-    }
-
-    pub(crate) fn ensure_group(
-        &mut self,
-        source_id: SourceId,
-    ) -> Result<&mut PartitionGroup<'cat>> {
-        let data_root = <_ as AsRef<Path>>::as_ref(&self.data_root);
-
-        Ok(self.groups.entry(source_id).or_insert_with(|| {
-            // this shouldn't fail in general
-
-            let root = PartitionGroupManager::new(data_root, source_id)
-                .with_context(|_| "Failed to create group manager")
-                .unwrap();
-
-            let mut pg = PartitionGroup::new(&root, source_id)
-                .with_context(|_| "Unable to create partition group")
-                .unwrap();
-
-            Catalog::create_single_partition(&mut pg);
-            pg.flush().unwrap();
-
-            pg
-        }))
-    }
-
-    /// Add new partition group with given source id
-
-    pub fn add_partition_group(&mut self, source_id: SourceId) -> Result<()> {
-        let _ = self.ensure_group(source_id)?;
-
-        Ok(())
-    }
-
-    fn create_single_partition(pg: &mut PartitionGroup) {
-        let part = pg.create_partition(MIN_TIMESTAMP)
-            .with_context(|_| "Unable to create partition")
-            .unwrap();
-
-        let mut vp = VecDeque::new();
-        vp.push_front(part);
-
-        pg.mutable_partitions = locked!(rw vp);
-    }
-
-
-    fn prepare_partition_groups<P, I>(root: P, ids: I) -> Result<PartitionGroupMap<'cat>>
-    where
-        P: AsRef<Path>,
-        I: IntoIterator<Item = SourceId>,
-    {
-        ids.into_iter()
-            .map(|source_id| {
-                let path = PartitionGroupManager::new(&root, source_id).with_context(|_| {
-                    format!(
-                        "Unable to obtain data directory for partition group {}",
-                        source_id
-                    )
-                })?;
-
-                let partition_group = PartitionGroup::with_data(path)
-                    .with_context(|_| format!("Unable to read partition group {:?}", source_id))?;
-
-                Ok((source_id, partition_group))
-            })
-            .collect()
-    }
-
-
-    fn serialize<P: AsRef<Path>>(catalog: &Catalog<'cat>, meta: P) -> Result<()> {
-        let meta = meta.as_ref();
-
-        let group_metas = Vec::from_iter(catalog.groups.keys());
-
-        let data = (catalog, group_metas);
-
-        serialize!(file meta, &data)
-            .with_context(|_| "Failed to serialize catalog metadata")
-            .map_err(|e| e.into())
-    }
-
-    fn deserialize<P: AsRef<Path>, R: AsRef<Path>>(meta: P, root: R) -> Result<Catalog<'cat>> {
-        let meta = meta.as_ref();
-
-        if !meta.exists() {
-            bail!("Cannot find catalog metadata {:?}", meta);
-        }
-
-        let (mut catalog, group_metas): (Catalog, Vec<SourceId>) = deserialize!(file meta)
-            .with_context(|_| "Failed to read catalog metadata")?;
-
-        catalog.groups = Catalog::prepare_partition_groups(&root, group_metas)
-            .with_context(|_| "Failed to read partition data")?;
-
-        catalog.data_root = root.as_ref().to_path_buf();
-
-        Ok(catalog)
-    }
-}
-
-impl<'cat> Drop for Catalog<'cat> {
-    fn drop(&mut self) {
-        self.flush()
-            .with_context(|_| "Failed to flush data during drop")
-            .unwrap();
-    }
-}
-
-impl<'cat> AsRef<ColumnMap> for Catalog<'cat> {
-    fn as_ref(&self) -> &ColumnMap {
-        &self.columns
-    }
-}
 
 #[cfg(test)]
 mod tests {
@@ -762,7 +50,9 @@ mod tests {
     const TIMESTAMP_SIZE: usize = 8; // should be `size_of::<Timestamp>()`
     const MAX_RECORDS: usize = BLOCK_SIZE / TIMESTAMP_SIZE;
 
-    fn create_random_partitions(pg: &mut PartitionGroup, im_count: usize, mut_count: usize) {
+    pub(super) fn create_random_partitions(pg: &mut PartitionGroup,
+                                           im_count: usize,
+                                           mut_count:usize) {
         let pts = RandomTimestampGen::pairs::<u64>(im_count + mut_count);
 
         let (imparts, mutparts): (Vec<_>, Vec<_>) = pts.iter()
@@ -802,7 +92,7 @@ mod tests {
 
             (init $columns: expr, $ts_min: expr) => {{
                 // assume 1, 5, 7 as default partition group ids for tests
-                use catalog::tests::append::DEFAULT_PARTITION_GROUPS;
+                use datastore::tests::append::DEFAULT_PARTITION_GROUPS;
 
                 append_test_impl!(init DEFAULT_PARTITION_GROUPS, $columns, $ts_min)
             }};
@@ -879,7 +169,7 @@ mod tests {
                     $data: expr,    // HashMap
                     $source_id: expr
                 ]),+ $(,)*) => {{
-                use catalog::tests::append::DEFAULT_PARTITION_GROUPS;
+                use datastore::tests::append::DEFAULT_PARTITION_GROUPS;
 
                 append_test_impl!(DEFAULT_PARTITION_GROUPS, $schema, $now, $([
                     $ts,
@@ -2006,66 +1296,6 @@ mod tests {
                     ]
                 );
             }
-        }
-    }
-
-    mod partition_meta {
-        use super::*;
-        use hyena_test::random::timestamp::{RandomTimestamp, RandomTimestampGen};
-
-        #[test]
-        fn deref() {
-            let id = Partition::gen_id();
-            let ts_min = RandomTimestampGen::random::<u64>();
-            let ts_max = RandomTimestampGen::random_from(ts_min);
-
-            let pm = PartitionMeta::new(id, ts_min, ts_max);
-
-            assert_eq!(*pm, id);
-        }
-    }
-
-    mod partition_group {
-        use super::*;
-
-        #[test]
-        fn new() {
-            let root = tempdir!();
-            let source_id = 5;
-
-            let pg = PartitionGroup::new(&root, source_id)
-                .with_context(|_| "Unable to create partition group")
-                .unwrap();
-
-            assert!(pg.immutable_partitions.is_empty());
-            assert!(acquire!(read pg.mutable_partitions).is_empty());
-            assert_eq!(pg.source_id, source_id);
-            assert_eq!(pg.data_root.as_path(), root.as_ref());
-        }
-
-        #[test]
-        fn with_data() {
-            let root = tempdir!();
-            let source_id = 5;
-            let im_part_count = 8;
-            let mut_part_count = 2;
-
-            {
-                let mut pg = PartitionGroup::new(&root, source_id)
-                    .with_context(|_| "Unable to create partition group")
-                    .unwrap();
-
-                create_random_partitions(&mut pg, im_part_count, mut_part_count);
-            }
-
-            let pg = PartitionGroup::with_data(&root)
-                .with_context(|_| "Failed to open partition group")
-                .unwrap();
-
-            assert!(pg.immutable_partitions.len() == im_part_count);
-            assert!(acquire!(read pg.mutable_partitions).len() == mut_part_count);
-            assert_eq!(pg.source_id, source_id);
-            assert_eq!(pg.data_root.as_path(), root.as_ref());
         }
     }
 
@@ -3512,49 +2742,6 @@ mod tests {
                 simple_sparse_u16, 3, 100_u16,
                 simple_sparse_u32, 4, 100_u32,
                 simple_sparse_u64, 5, 100_u64,);
-        }
-    }
-
-    mod catalog {
-        use super::*;
-
-        #[test]
-        fn new() {
-            let source_ids = [1, 5, 7];
-            let im_part_count = 8;
-            let mut_part_count = 2;
-
-            let root = tempdir!();
-
-            let mut cat = Catalog::new(&root)
-                .with_context(|_| "Unable to create catalog")
-                .unwrap();
-
-            for source_id in &source_ids {
-
-                let pg = cat.ensure_group(*source_id)
-                    .with_context(|_| "Unable to retrieve partition group")
-                    .unwrap();
-
-                create_random_partitions(pg, im_part_count, mut_part_count);
-            }
-        }
-
-        #[test]
-        fn add_partition_group_idempotence() {
-            let root = tempdir!();
-
-            let mut cat = Catalog::new(&root)
-                .with_context(|_| "Unable to create catalog")
-                .unwrap();
-
-            const PG_ID: SourceId = 10;
-
-            cat.add_partition_group(PG_ID).unwrap();
-            cat.add_partition_group(PG_ID).unwrap();
-
-            assert_eq!(cat.groups.len(), 1);
-            assert_eq!(cat.groups.iter().nth(0).expect("partition group not found").0, &PG_ID);
         }
     }
 }
